old/data/requirements.txt
Pillow>=9.0.0
pillow-heif>=0.7.0 